+++ backend/tests/sum.test.ts
@@ -1,5 +1,5 @@
 import {describe, expect, test} from '@jest/globals';
-import {sum} from './sum';
+import {sum} from '../src/sum';
 
 describe('sum module', () => {
   test('adds 1 + 2 to equal 3', () => {
